review comments pass #1.1
