+++ stats/opentelemetry/csm/pluginoption_test.go
@@ -21,7 +21,6 @@ package csm
 import (
 	"context"
 	"encoding/base64"
-	"encoding/json"
 	"fmt"
 	"net/url"
 	"os"
@@ -31,7 +30,7 @@ import (
 	"google.golang.org/grpc/internal/envconfig"
 	"google.golang.org/grpc/internal/grpctest"
 	"google.golang.org/grpc/internal/testutils"
-	"google.golang.org/grpc/internal/xds/bootstrap"
+	"google.golang.org/grpc/internal/testutils/xds/e2e"
 	"google.golang.org/grpc/metadata"
 
 	"github.com/google/go-cmp/cmp"
@@ -335,16 +334,7 @@ func (s) TestBootstrap(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			bootstrapContents, err := bootstrap.NewContentsForTesting(bootstrap.ConfigOptionsForTesting{
-				Servers: []json.RawMessage{[]byte(`{
-					"server_uri": "xds_server_uri",
-					"channel_creds": [{"type": "insecure"}]
-				}`)},
-				NodeID: test.nodeID,
-			})
-			if err != nil {
-				t.Fatalf("Failed to create bootstrap configuration: %v", err)
-			}
+			bootstrapContents := e2e.DefaultBootstrapContents(t, test.nodeID, "xds_server_uri")
 			testutils.CreateBootstrapFileForTesting(t, bootstrapContents)
 			nodeIDGot := getNodeID() // this should return the node ID plumbed into bootstrap above
 			if nodeIDGot != test.nodeID {
@@ -491,16 +481,7 @@ func (s) TestSetLabels(t *testing.T) {
 					defer os.Unsetenv("CSM_WORKLOAD_NAME")
 				}
 				if test.bootstrapGeneratorPopulated {
-					bootstrapContents, err := bootstrap.NewContentsForTesting(bootstrap.ConfigOptionsForTesting{
-						Servers: []json.RawMessage{[]byte(`{
-							"server_uri": "xds_server_uri",
-							"channel_creds": [{"type": "insecure"}]
-						}`)},
-						NodeID: "projects/12345/networks/mesh:mesh_id/nodes/aaaa-aaaa-aaaa-aaaa",
-					})
-					if err != nil {
-						t.Fatalf("Failed to create bootstrap configuration: %v", err)
-					}
+					bootstrapContents := e2e.DefaultBootstrapContents(t, "projects/12345/networks/mesh:mesh_id/nodes/aaaa-aaaa-aaaa-aaaa", "xds_server_uri")
 					testutils.CreateBootstrapFileForTesting(t, bootstrapContents)
 				}
 				var attributes []attribute.KeyValue
